Rename AbstractAIResource to AIResource to match naming convention (#461)
* Rename AbstractAIResource to AIResource to match naming convention

All other classes which define the (database) relationships are
just named after their conceptual model's counterpart, e.g.,
AIoDConcept, AIAsset, Publication. This seems to be the exception
with no apparent reason. Changing the name to match the convention
of the rest of the project should make it easier to understand
the role of the class.

* Merge import statements